WT-2554 Fix build warnings (#2740)

+++ test/utility/parse_opts.c
@@ -120,12 +120,12 @@ testutil_parse_opts(int argc, char *argv[], TEST_OPTS *opts)
 	 * Setup the home directory. It needs to be unique for every test
 	 * or the auto make parallel tester gets upset.
 	 */
-	len = snprintf(NULL, 0, "WT_TEST.%s", opts->progname) + 1;
+	len = (size_t)snprintf(NULL, 0, "WT_TEST.%s", opts->progname) + 1;
 	opts->home = (char *)malloc(len);
 	snprintf(opts->home, len, "WT_TEST.%s", opts->progname);
 
 	/* Setup the default URI string */
-	len = snprintf(NULL, 0, "table:%s", opts->progname) + 1;
+	len = (size_t)snprintf(NULL, 0, "table:%s", opts->progname) + 1;
 	opts->uri = (char *)malloc(len);
 	snprintf(opts->uri, len, "table:%s", opts->progname);
